bug #102728: RFE: Develop Oracle ACS compatible executables for rmcd
This commit does two things.  Firstly it adds the first version of the source
code for the executables to help the rmcd daemon to work with Oracle ACS
compatible tape-libraries.  Secondly it refactors and cleans up some of the
rmcd daemon code.

Please note that the source code of the binaries to help the rmcd work with
Oracle ACS tape-libraries is neither compiled or packaged into rpms.  Therefore
there will be no affect on the current 2.1.14 release concerning these
executables.