bug #102853: RFE: Remove code implementing unsupported tape label types
The refactoring carried out for this RFE has uncovered what maybe a potential
weakness in the code used to flush data to tape.  This modification is based on
what we have learnt about the legacy code during its refactoring.  It improves
the safety of the tape flush logic.  The current code in production has never
failed, however it is always better to be safer than sorry.