I have moved all of the logrotate configuration scripts to the directories that contain the Imakefiles that work with them.
The logrotate configuration scripts were in the CASTOR/debian directory for a
legacy reason.  In the past there was only one configuration for all of CASTOR
and therefore a central directory such as the debian directory made sense.
With the advent of the /etc/logrotate.d directory each package/server could
have its own individual lograte configuration file.