castor/tape/tapegateway/CMakeLists.txt now builds all of it directory.
Please note that I have explicitly not ported the installing of the
tapegateway header files.  These headers files are currently installed by the
castor-build-headers rpm.  I beleive that they should not be part of the
castor-build-headers rpm in 2.1.15-x because the new tape server code will live
within the main CASTOR git repository and will therefore not need to have an
rpm such as castor-build-headers in order to construct a build environment
outside of CASTOR.