Fixed bug #103370: The logic to resume recall jobs after an unmount is broken when dealing with double copy recalls.
This fix is not optimal: we're still using fileTransactionId being NULL
as a flag while we should really create an extra status
RECALLJOB_SELECTED2ND or SELECTEDSHADOW. Not on this hot fix...

This commit also includes a simpler way to select new recall candidates,
which makes it more immune to bugs like the above.
Plus a fix related to the precision used to check if a file was
overwritten while being recalled.