After discussions with Eric I have removed the singleton pattern from the implementation of the proposal for a new CASTOR logging system API.
The new API provides a Log class which contains a mutex that mirrors the
previous locking strategy of Sebastien and Denis.  My additional singleton mutex
and functionality should not be in the logging API.  Giuseppe and Sebastien
should decide how the logging API is exposed to subclasses of CASTOR classes
such as Server and Daemon.  The loggin API should concentrate on logging.

By the way, instantiating two Log objects is not actually a problem, they will
simply have separate connections to syslog, that's all.