bug #103642: Remove the deltpfil() function
Removed the deltpfil() function that deletes files from the end of a tape in
order to leave the tape terminated with a correct label structure.  The
deltpfil() function increases the risk of data loss in a system that uses
deferred tape flushes.  When appending data to the end of a tape the
deltpfil() function has been made redundant by the fact that the stageri
always instructs the tape system to position to the end of the last file
successfully written to tape as opposed to requesting a tape drive to position
to end of data.