bug #103642: Remove the deltpfil() function
The current removal of the deltpfil() function is not satificatory.  The current
solution effectively replaces the deltpfil() function with a log message and a
call to exit(-1).  The call to exit(-1) prohibits the propagation of a TAPE
OVERFLOW error to the tapegatewayd daemon in the event that a tape server
writes past the physical end of a tape.  This side effect means that tapes are
mounted in a loop once they become physically FULL because the vmgrd daemon is
never notified that they are FULL.  The mount loop is actually finite because
depending on the length of the last file(s) be written to tape the TAPE OVERFLOW
is detected by different parts of the tape server code and the parts that would
not have called the deltpfil() function are able to propagate the TAPE OVERFLOW
error to the tapegewayd daemon which in turn is able to notify the vmgrd daemon
that the tape in question is now FULL.

The new idea that is yet to be implemented, is to simply log a message instead
of calling deltpfil() and then let the rtcpd tape session end gracefully.  This
graceful end of tape session must not run the risk of trying to write to tape
again in some form of retiry logic.  I am therefore currently trying to remove
all retry related code.  One such collection of related code is the now
obsolete abort protocol, where a client of rtcpd sends it an RTCP_ABORT_REQ
message.  This commit removes the defunct marshalling code for RTCP_ABORT_REQ
messages from the tapebridged daemon.