Revert "Removed the following compiler flag from the tape directory of CASTOR:"
This reverts commit 308b14d257161b42a78ac0972b0f17cfb530b329.

From leaning on TeamCity, I see that the following flag is needed for SLC6:

    -fno-strict-aliasing

This is really annoying because I cannot see these SLC6 compilation errors on
my development box because it's SLC5.  The effort needed to spot and fix all of
the -O2 bugs is increased by the fact that SLC5 and SLC6 do not give the same
errors.