Fixed bug #103659: RFE: stager_get should warn/error on unknown command line arguments.
Actually all command lines were affected, and in particular
stager_putdone used to work by 'mistake': both the command line
parsing and the output did not respect the man page and the API.
Now this is fixed and the change reflected in the test suite.
Moreover, a new test has been introduced, which checks some of the
possible invalid command line arguments.