bug #103798: RFE: Add dry run to recall mount logic to help prevent empty recall mounts
This reverts commit 5bdc2b3399c949e29d770ba0f32739d021e00066.

This is in fact the revert of the revert (now under which cup is the coin?).

I am following the suggestion of Giuseppe that in order not to lose Eric's
work it would be a good idea to put his fix for his RFE into the master branch
as soon as possible.