I have made m_foreground and m_commandLineHasBeenParsed private member variables of the following class:
    castor::server::Daemon

Subclasses now have a more explicit API for parsing the command-line.  They
can delegate the task to the above Daemon class or they can implement there own
parsing logic.  In the latter case the subclass must call the following
Daemon method:

    castor::server::Daemon::setCommandLineParsed(bool foreground);

This method makes it clear what the Daemon class needs to know from a parse of
the command-line.  If a client subclass calls the getForeground() method of the
Daemon class before callng setCommandLineParsed() then a CommandLineNotParsed
exception shall be raised.