Replaced castor::io::marshallValue() by type specific marshall functions.
The castor::io::marshallValue() function was incorrectly named.  The function
was meant to be used soley for marshalling integer types and not any type.  The
function did not use the htonl() and htons() where they could have been used.
CASTOR has to marshal 64-bit numbers and in this case it is true that the
htonl() family of functions is of no use, however for 16-bit and 32-bit numbers
we should use them.