Replaced castor::io::unmarshallValue() by type specific marshall functions.
The castor::io::unmarshallValue() function was incorrectly named.  The function
was meant to be used soley for unmarshalling integer types and not any type.
The function did not use ntohl() and ntohs() where they could have been used.
CASTOR has to unmarshal 64-bit numbers and in this case it is true that the
ntohl() family of functions is of no use, however for 16-bit and 32-bit numbers
we should use them.