I have fixed castor::io::marshalUnit64().
The castor::io::marshalUint64() was not marshalling a 64-bit integer in the
same was as the legacy code of h/marshall.h.

The legacy code was treating a 64-bit integer as two 32-bit integers. It
would first marshall a 32-bit iinteger representing the lower powers of 2
and then marshall a 32-bit integer representing the higher powers of 2.

The castor::io::marshalUint64() method was incorrectly treating a 64-bit
integer as a whole by simply marshalling the whole 8 byte block of
memory in reverse byte order.