Added support for VIRTUAL tape drives, for tests
We now support a new "model" of drives called virtual. This will allow
auto-detection of drives in fake environement, for unit tests.
Changed accordingly the signature of the reference returned by the Drive
wrapper from DriveGeneric to DriveInterface
Also fixed the order of the parameters of setDensityCode, to allow the usage
of the default densityCode.