Exception class Errnum and SErrnum now call sstrerror()
The sstrerror() function was developed in CASTOR to deal with SHIFT
error codes as well as standard errno error codes.  The sstrerror()
was also supposed to be thread safe.  I believe that this is not the
case and if I am right then it will be fixed for all of CASTOR in a
later commit.