Added the strerror_r_wrapper()
The  strerror_r_wrapper() function wraps the XSI compliant version of
strerror_r().

This wrapper facilitates the task of having the XSI compliant verson
of strerror_r() as opposed to the GNU version.  In order to the XSI
compliant version, the implemenation file common/strerror_r_wrapper.cpp
undefines _GNU_SOURCE and defines _XOPEN_SOURCE to be 600.

Please note that the strerror_r_wrapper() is not yet called by any
other code in CASTOR.  The purpose of this commit is to see how the
wrapper function compiles on SLC5, SLC6 and MAC.