Added the private method VdqmImpl::connectToVdqm()
This version of connectToVdqm() is based on Cgethostbyname().  This is
a temporary implementation.  I would like that it is replaced by code
based around getaddrinfo() which is both standard and thread-safe on
linux and is therefore better than gethostbyname() whose rentrant
cousin gethostbyname_r() is a GNU extension.  According to the BSD
manual pages, the getaddrinfo() function is a replacement for and
provides more flexibility than gethostbyname().

I cannot find any information about the thread safety of getaddrinfo()
on MAC.