Fixed a bug in sstrerror()
I introduced a bug into sstrerror() when I added error checking code
around the call to Cglobals_get().  The hash of the offending commit is:

051afce974bb2916851e46394e1b1f562e3b776a

Before my so say improvements, the return value of Cglobals_get() was
not checked.  I added a check that assumed an error had occured if the
return value was non-zero.  Unfortuntaley for me, the return values 0
and 1 both mean success.  A return value of -1 signifies a failure.