Provided hook method TapeDaemon::notifyLabelCmdOfEndOfSession()
The labeling of a tape now requires TapeDaemon to have two different types
types of session: data transfer sessions and label sessions.  The reaper
logic of the TapeDaemon parent process has to deal with the end of these
two types of session differently.  In the case of a label session the
TapeDaemon parent process must notify the client of the success or failure
of the label session.

Labeling a tape also requires TapeDaemon to have its first long running
and synchronous request.  Labeling a tape includes mounting and unmounting
that tape which can take several minutes.  The client command-line tool
for labeling tapes should be put on hold whilst the full process of
labeling takes place.