Exceptions thrown by I/O reactor of tapserverd are now logged and skipped
With our current understanding we see no reason for an exception from the
reactor to be used as a reason to stop the tapeserverd daemon.

In the future, one or more specific exception types could be introduced
with their own catch clauses that do in fact require the tapeserverd daemon
to be stopped.