Added commented out "castor_tape_reactor_ZMQReactorTest, closeFd"
The destructor of the ZMQPollEventHandler class is currently not
virtual when it should be.  It should be virtual so that
subclasses can close the file descriptors that they own.  It is
not virtual because the new tape server software freezes due to a
strange interaction with ZMQ which is currently being investigated
and hopefully fixed.

When the destructor of the ZMQPollEventHandler class is finally
made virtual the following test should be uncommented:

    File: castor/tape/reactor/ZMQReactorTest.cpp
    Test: TEST_F(castor_tape_reactor_ZMQReactorTest, closeFd)
    Valrgind command-line:
        valgrind --track-fds=yes test/castorUnitTests