Fixed xroot OFS plugin for ceph support.
In principle, this support was supposed to be provided by an independent OSS plugin, but it happens that external attributes are not supported by OSS and that POSIX calls were used directly, bypassing OSS. Thus a wrapper had to be used for the moment around the POSIX calls. To be imprived if OSS supports external attributes one day.