Fixed bug where putting down a running drive crashes tapeserverd
Vlado found this bug whilst running tapserverd with real tape hardware.
Mounts last long enough with real hardware for an operator to easily
use "tpconfig down" on a running drive.

The tapserverd process was not handling the WAIT_DOWN state correctly
and was throwing an unexpected exception that in turn caused the process
to crash due to the resulting abort signal.

The tapserverd process was not handling the WAIT_DOWN state correctly
due to some over simplifications I made to the architecture of
tapeserverd.