Removed false/emulated nameserver check from LabelSession
It was planned to have the LabelSession of the new tapeserverd
daemon check that a tape being labeled was empty with respect to
the CASTOR namespace.  This feature, which currently does not
exist within either rtcpd or taped has turned out to be more
difficult to implement than expected.  This new additional feature
is therefore being dropped before it gets released.