CASTOR-4739: tapeserverd should support localfile, rfio, xroot and rados striper access for disk files
After testing Xroot access, realised that we need to add a signature in the opaque data of the
Xroot URL. A signature using OpenSSL, as used in other places of the project was implemented,
but did not pass the memory leak and race condition tests. As discussed with the rest of the team,
OpenSSL is not to be trusted and has poor quality memory managment. The CryptoPP library, also part
of the SLC6 distribution has been tested as well, and it output validated (on a few example) with
OpenSSL's. CryptoPP can be used instead of OpenSSL, and has been put in place. This commit still contains
the OpenSSL for reference. The next commit will remove it.