CASTOR-4739: tapeserverd should support localfile, rfio, xroot and rados striper access for disk files
Changed the behaviour in case of non-URL paths(*): if we get Ceph non-URL path, we open
the path (which always points to localhost) via the castor2fs xroot extension. In this
case, we add the castor2fs.pool parameter to the path. We also force to port number
to castor2fs's 1095. This will allow access to ceph via  local xroot translation deamon
running on the tape server's host.

The code has been factored out to allow also bare Xroot accesses with full URLs.

(*) Non-URL path is the current path description exchanged between tape gateway and
tape server, in the form <server>:<pool>/<object name> for ceph objects and
<server>:/path/to/the/file for xroot files.