CASTOR-4823: tapeserverd's main process should log statistics and status of the closing session in all cases
Added new messages between the session process and the main process: addition and removal of log parameters.
Added storage of the log parameters in the catalogueTransferSession.
Added logging of the tape session at the end in the initial process, allowing statistics gathering even in
case of crash. This standard log message is intended to be used as an input for the tape log system.
Changed the default off the formatting of doubles in the log so that we don't end up with hard to parse
scientific notation in the logs.