tapeserverd now synchronises the vdqm with the status of its drives
The VDQM protcol is fragile.  A single network or database error (for
example a translucent upgrade) can put the vdqm out of sync with the
tape drives it controls.  This modification makes tapeserverd check
with the vdqm from time to time to see if everything is still
synchronised.  If it is not the case then tapeserverd now takes
proactive action to resync the vdqm.