Added Oracle error ORA-01008 'not all variables bound' to the list of Oracle error codes that shall trigger a reconnection.
This error is not properly a 'lost connection' error, but we have
repeatedly observed cases where the DB connection gets corrupted
and the server spits out this error message for perfectly valid
statements (last occurrence being an incident in castorpublic this
morning). A reconnection is a valid workaround to get back in
business.