XROOTD: Add option to skip aborting orphan requests
The new configuration directive "xcastor2.skipabort" is by default false and
can be enabled by setting it to true. If this is enabled then we no longer
abort the in-flight requests for which the client disconncted i.e. we no longer
issue stage abort requests.

Note: This is put in place to accommodate the Alice workflow. The proper way of
achieving this on the client side would be to do a prepare request before trying
to open the file.