Scheduler now calls SchedulerDatabase::queue(const ArchiveToDirRequest &).
Before this commit the Scheduler class would call

  SchedulerDatabase::queue(const ArchiveToFileRequest &)

in a loop when handling a user request to archive multiple remote files
to a directory in the archive.  This modification opens the door for the
SchedulerDatabase object to remember more about the context of each
archive request.  In this case which archive to directory request was at
the origin of each of the corresponding archive to file requests.