XROOTD: Add support for multiple public keys on diskservers
The diskserver can hold now multiple public keys that are used to
verify the signature of the requested transfers. The public keys
are saved in a map holding info about the key id and the key location
on the local filesystem. There needs to be at least one "default"
key type which is used if no other key matches the type of transfer.
The rest of the keys should have an id that matches the type of
transfer. Possible types of transfes: user/tape/d2duser/tape/d2duser/
d2dbalance/d2dinternal/d2ddrain. The transfer type is used to select
the proper public key to verify the signature.