Added SqliteCatalogueSchema so IDEs work with SqliteCatalogue
The CMakeLists.txt file of the catalogue directory instructs
cmake to generate SqliteCatalogueSchema.cpp, by combining
the contents of the SqliteCatalogueSchema.before_SQL.cpp with the
contents of catalogue_schema.sql.  This means the
SqliteCatalogueSchema.before_SQL.cpp file is not compilable.

The purpose of the SqliteCatalogueSchema structure class is to
isolate the "non-compilable" issues into a small and encapsulated
compilation unit, namely SqliteCatalogueSchema.o, so that the
remaining business logic can be implemented in the non-generated
and compilable SqliteCatalogue.cpp file.  This means that IDEs can
work as normal with SqliteCatalogue.cpp, whereas they will struggle
with SqliteCatalogueSchema.before_SQL.cpp which is therefore
intentionally small.