Replaced select() with poll() in castor::io
This feature is the result of porting the following CASTOR commit:

    commit 866e31e008b8d8eb5068ad0a887685cd6a1a9eb5
    Author: Steven Murray <steven.murray@cern.ch>
    Date:   Wed Jul 20 11:06:38 2016 +0200

        Replaced select() with poll() in castor::io

The use of select() in the CASTOR tapeserverd daemon prevented the
daemon from working directly with Ceph.  Ceph creates thousands of
concurrent TCP/IP connections which means the "next" file
descriptor to be used for say a new connection from tapeserverd to
rmcd would have a value larger than the maximum value supported by
select().