Simplified the interface of the TapedProxy and TapeServerReporter.
The TapeServerReporter, which is the main interface for the tape thread and tasks,
now has a simple reportState() function instead of mnay ad-hoc ones. The only ones
remaining are reportTapeUnmountedForRetrieve() and reportDiskCompleteForRetrieve()
as they allow managing the special case of retrieve where the session can from
Running to either ShutingDown or DrainingToDisk depending on which order the
threads complete.

The actual calls to send messages to taped are now 3: reportState,
addLog, removeLog.