Ported commit fab791bc805f95d59c55cbcb554e6f218710b0dd from castor/master
CASTOR-4836: tapeserverd should have a new time counter: delivery time

Fixed.

Add two counters: deliveryTime and drainingTime. For the recall
session deliveryTime is the total time of the disk threads and
drainigTime is the time difference between deliveryTime and
the total time of the tape thread or in other words how much time
disk threads spent after the tape thread finished. For the
migration session deliveryTime is the total time of the tape thread
and drainingTime always equal 0.