Implement RAO at the drive level #57
This commit adds the feature of querying the drive for the
Recommended Access Order (RAO) of a series of files and adds
a test file for this feature.

Changes of this commit:
  *  Extend the DriveInterface with a virtual method queryRAO
which takes as parameter the name of the file containing a set
of files for which the RAO is requested. The line format of this
file is the following:

FILE_ID:BLOCK_START:BLOCK_END

  *  Extend DriveGeneric class with the implementation of the
method queryRAO, along with a few other auxiliary methods; some
key auxiliary methods are getLimitUDS, generateRAO, receiveRAO

  *  Add method queryRAO in object FakeDrive returning Exception:
Not Implemented

  *  Add auxiliary constants and structures in a new 'RAO' namespace

  *  Add test file: BasicReadWriteTest.cpp which can perform 3 tests
according to the setting of variable 'test'
	1. BLOCK_TEST - basic read / write of blocks
	2. FILE_TEST - writes a series of CASTOR files and reads
a random one
	3. RAO_TEST - performs a RAO query for a series of files
given in the filename specified in argv[1]. The file format should
respect the above mentioned order.

Signed-off-by: Cristina Moraru <cristina-gabriela.moraru@cern.ch>