Revert "Removed 'Do nothing' logic from cta::server::Daemon which unwantedly kicks in if systemd is used"
This reverts commit 462959d43abe211937da8d0f51d4216f78aefb57.

The current pipeline in the GitLab CI keeps failing to complete
the "init" pod's work.  The same Docker image is used on
stevedevctai.cern.ch to successfully run systest.sh -n ctatest.

The GitLab CI does not provide any logs from the "init" pod and
therefore cannot be debugged.

This revert is a desperate attempt at fixing ithe problem.  If it
does work then it means the GitLab CI is benefiting from cta-taped
remaining as the root user.