cta/CTA#121 cta tape add requires optional --encryptionkey parameter
I have updated the help page of the "cta tape" command.  The page
now explains that encryption_key is either "none" or the name of the
individual key to be used.

Please note that encryption and its configuration is not for release
0 of CTA.  It is also not clear how much encryption configuration
details should be inside of the CTA catalogue.  We may for example
drop the concept of encryption_key name all together and just record
a boolean that simply indicates whether or not a tape is encrypted.