cta/CTA#106 cta-taped logging a timestamp
All sub-classes of the Logger class, excluding DummyLogger and
SysLogLogger, now prefix each log message with the same header
as CASTOR.

Please note that SysLogLogger class intentionally leaves the
creation of log message headers to rsyslog.  It is the
responsibility of the system administrator/operator to configure
rsyslog to generate the header syntax they wish including the
precision of the timestamp within the header.