cta/CTA#139 cta af ls --vid XXXXX command times out if there are many files on a tape
Added the XrdCtaFile::m_listingArchiveFiles member variable and
factorized XrdCtaFile::readFromCmdlineOutput() out from
XrdCtaFile::read() so that in the future the XrdCtaFile::read()
method will be able to treat the listing of archive files
differently from other cta commands.

Currently none of the cta commands are incrementally generating
and streaming large result sets back to the cta command-line
tool.  All commands are generating the entire result in the
constructor of XrdCtaFile.  The result is streamed back to the
cta command-line tool, but the result is generated in one
single stepi before any streaming starts.

The listing of archive files is being gradually modified to
incrementaly generate the result in each call to
XrdCtaFile::read().  This commit is one step in this
larger modification.