Increased the archive queue update delay.
It appeared in production that the update latency could be steadily higher than this delay.
This led to an ever growing perceived latency by the client, as the batches to be queued accumulated
faster than they were drained.

This is an interim solution to a more adaptive one which should make no asumption.