[EOS-CTA] Integrates XrdSsi Service object
The framework now expects only 3 protobuf types (Request, Metadata-response
and Alert). This is because the response payload will be arbitrary data of
arbitrary length and so is less well suited for protocol buffers.

We can define the characteristics of the response (size, format, etc.) in
the Metadata protobuf, or (if there is a use case), we can split the response
into records and use a protobuf for the record format.