Dropped the TAPE_FILE_VID_ARCH_FILE_ID_UN constraint
This constraint prevents a tape server from writing the same file
more than once to the same tape.  We should actually allow this
to happen so that we can tolerate the pathological case of a
cta-taped daemon crashing immediately after inserting a row into
the TAPE_FILE table and before updating the object store.