Simplified the AgentReference ownership operations queueing.
The queue is now fully event driven: a write to object store will start as soon as the previous one completed (if any).
The notion of timeout and max number of elements before flush has been removed. This harmonises the mechanism with the
one used in MemArchiveQueue::sharedAddToArchiveQueue() (see a591129c).