xCom_deletearchive does not call the object store
This is a TEMPORARY modification to the  CTA front end.

When an "eos rm" command is executed the xCom_deletearchive()
method of the CTA front is called.  Before this commit the
mthod would try to removed the archive file from both the
object store and the CTA catalogue.  This took more than 15
seconds to complete under som ecircumstances.  This commit
removes the call to the object store.

Please note that this commit will causes a leak in archive
files.  If a file is "in-flight" and therefore has not yet
been written to tape and recorded in the CTA catalogue
then a call to "eos rm" will do nothing and the file will
eventually appear in the CTA catalogue when it is written
to tape.  The file will therefore NOT be in the EOS
namespace but it WILL be in the CTA catalogue.