Added after requeueing requests in {Archive,Retrieve}Request::garbageCollect().
As those fuctions are usually called in a tight loop at the end of the session or in the garbage collector,
they tend to unlock and relock the queues immediately, and sitting on the lock for a long time in practice.
The added sleep will increase the odds of other processing slipping in between the unlock and the lock.