[cta_frontend] Refactors dispatcher to use a map of functions
For the admin functions, it is very clunky to map <cmd,subcmd> pairs to
which function to call using nested switches. Instead we lookup which
function to call in a static const map of <cmd,subcmd> to method function
pointers. Now new commands can be added or old ones removed by editing a
single line in the map definition.