Fixed long updating of statistics in object store leading to timeout.
The updating of statistic in the objecttore (in drive register) could wait for a lock
for so long that the session get killed due to a heatbeat timeout.

To avoid this, we use lockfree access as much as possible, and when locking it srequired, do it with a
timeout.

The timeout will be logged as a warning.

Added timeouts to synchronous locking primitives in the object store backends to implement this feature.

Fixed a bug where the VFS backend always used exclusive locks.