Implement 'cta drive rm' command.
Command will remove drive or drives by regex from the drive register.
The drives must be in state Down unless --force option is set.
Usage: rm <drive_name> [--force/-f]

If there is any active tapeserver with this drive it will create the
entry in the drive register again with Down status.

Example of the usage:
  cta drive rm ^VDSTK
  Drive VDSTK41 in state Free and force is not set (skipped).
  Drive VDSTK42 removed.