cta af ls -all no longer sorts the resuls
If the whole contents of the ARCHIVE_FILE and TAPE_FILE tables are
listed without a search criteria and with an ORDER BY clause then the
first row back from the database will take a relatively long period of
time to be returned.  This is because the database will have to sort
every row before returning the first.  This long period of time will
cause the current (non SSI) cta command-line tool to time out.