[cta_frontend] Integrates cta_admin af ls command
This is a first version of the command which has a fixed-size return buffer,
this code still needs to be refactored to allow a result of arbitrary size.

Removes the log parameter from ListArchiveFilesCmd constructor. Elsewhere we
are using log context objects: the SSI version of the command dispatcher has
access to the lc but not the log object. In any case this was not being used
so safe to remove for the time being.